transfer float to double precision

+++ include/cufht.h
@@ -3,13 +3,13 @@
 #include <vector>
 
 // using DataType = unsigned __int8;
-// using DataType = float;
+// using DataType = double;
 
-void initializeKernelHost(std::vector<float>& kernel, const int cols);
+void initializeKernelHost(std::vector<double>& kernel, const int cols);
 /**
-* DHT1DCuda(float* vector, const int length) returns the Hartley
+* DHT1DCuda(double* vector, const int length) returns the Hartley
 * transform of an 1D array using a matrix x vector multiplication.
 */
-void DHT1DCuda(float* vector, const int length);
+void DHT1DCuda(double* vector, const int length);
 
-void HT2DCuda(const std::vector<float>& X, std::vector<float>& Y, const int cols, const int image_num);
+void HT2DCuda(const std::vector<double>& X, std::vector<double>& Y, const int cols, const int image_num);
