+++ include/kernel.h
@@ -2,9 +2,9 @@
 #define KERNEL_CUH_
 
 //using DataType = unsigned __int8;
-//using DataType = float;
+//using DataType = double;
 
-void matrixMultiplication(float *A, float *B, float *C, int N);
-void vectorMatrixMultiplication(float* A, float* B, float* C, int N);
+void matrixMultiplication(double *A, double *B, double *C, int N);
+void vectorMatrixMultiplication(double* A, double* B, double* C, int N);
 
 #endif
